working shooter example

+++ src/mugato/action/Action.hpp
@@ -2,6 +2,8 @@
 #ifndef __mugato__Action__
 #define __mugato__Action__
 
+#include <memory>
+
 namespace mugato {
 
     class Entity;
@@ -11,7 +13,7 @@ namespace mugato {
     public:
         virtual ~Action(){}
         virtual void start(Entity& entity){};
-        virtual void update(Entity& entity, float pos){};
+        virtual void update(Entity& entity, double pos){};
         virtual void finish(Entity& entity){};
     };
 }
